update Map probe
update the mock system of maping

+++ Delivery_Service/Database/schema.sql
@@ -17,8 +17,8 @@ name varchar(255),
 averagePay int,
 averageTime int,
 numberOfDelivery int,
-longitude int,
-latitude int,
+longitude float,
+latitude float,
 lastUpdate DATETIME
 );
